Add simulation modules. Update to version 1.0.
Here we add all of the simulation modules, which are all of the *.py
library files in the top directory. We have also included the research
notebook as well. The simulation data in the *_runs folder will not be
included in the git repository, but will stay in the Dropbox directory
for backup purposes. We have not added any plotting scripts or images.

This is the "version 1.0," as we have started making
profession plots with the data so far.