Update to "runsim symlink" run directory structure.
For automatic documentation/SSOT design.  Each run directory is now
self contained in a directory named "series", where we have defined
series="whatever descriptive word that we would like to name our
figures after". No longer copy pasting the first 100 lines of code
from a "test-*.py" into plotting applications.  This will be standard
going forward, but "apsidal-alignment" project will remain backwards
compatible because this update simply splits the original test-*.py
files into two files and hooks them together in the following way:

../project/*symlinked-runfile-dir*/
	series-run.py
	series-params.py

../project/
	runsim -> series/series-run.py

Then, executing "python runsim" in ../project/ directory will
check if ../project/series/ exists, and, if not, it will
create it. Next, it enters that directory and begins executing
simulations. All of this happens irrespective of the value
of *symlinked-runfile-dir*, so that the naming scheme
"series-run.py" and "series-params.py" automatically
create output in a directory named "../project/series/".

*Note:* could just have chdired and "python series-run.py",
but it's helpful to manually change the symlink when
switching contexts.